1768 update wallet tab (#2256)
* wip: #1798 changet TokenHolder layout, added large title, added search (wip), moved add / hide token to the very bottom
* wip: #1798 implemented search
* wip added absolute token price change to token holder
* wip: added header holder
* WalletConnect chain choice (#1999)
* Allow selection of WalletConnect chain in the session dialog
* Apply modified heuristic for token names and symbols
* added groups (Assets / NFT) to the token adapter
* updated text filter on tokens adapter #2001
* Refine search update
* added read of the aTokens list #1768
* added wallet total #1768
* added wallet value change for today #1768
* Enlarge the font for 24H Price change #2021
* fixed formatting issue for the percents 24h change
* implemented design tweaks as requested in #2230 (#2250)
* Add summary sync and total value

+++ app/src/main/java/com/alphawallet/app/repository/TransactionRepositoryType.java
@@ -18,23 +18,23 @@ import io.reactivex.Single;
 import io.realm.Realm;
 
 public interface TransactionRepositoryType {
-	Single<String> createTransaction(Wallet from, String toAddress, BigInteger subunitAmount, BigInteger gasPrice, BigInteger gasLimit, byte[] data, int chainId);
-	Single<TransactionData> createTransactionWithSig(Wallet from, String toAddress, BigInteger subunitAmount, BigInteger gasPrice, BigInteger gasLimit, long nonce, byte[] data, int chainId);
-	Single<TransactionData> createTransactionWithSig(Wallet from, BigInteger gasPrice, BigInteger gasLimit, String data, int chainId);
-	Single<TransactionData> getSignatureForTransaction(Wallet wallet, Web3Transaction w3tx, int chainId);
-	Single<SignatureFromKey> getSignature(Wallet wallet, Signable message, int chainId);
-	Single<byte[]> getSignatureFast(Wallet wallet, String password, byte[] message, int chainId);
+	Single<String> createTransaction(Wallet from, String toAddress, BigInteger subunitAmount, BigInteger gasPrice, BigInteger gasLimit, byte[] data, long chainId);
+	Single<TransactionData> createTransactionWithSig(Wallet from, String toAddress, BigInteger subunitAmount, BigInteger gasPrice, BigInteger gasLimit, long nonce, byte[] data, long chainId);
+	Single<TransactionData> createTransactionWithSig(Wallet from, BigInteger gasPrice, BigInteger gasLimit, String data, long chainId);
+	Single<TransactionData> getSignatureForTransaction(Wallet wallet, Web3Transaction w3tx, long chainId);
+	Single<SignatureFromKey> getSignature(Wallet wallet, Signable message, long chainId);
+	Single<byte[]> getSignatureFast(Wallet wallet, String password, byte[] message, long chainId);
 
     Transaction fetchCachedTransaction(String walletAddr, String hash);
 	long fetchTxCompletionTime(String walletAddr, String hash);
 
-	Single<String> resendTransaction(Wallet from, String to, BigInteger subunitAmount, BigInteger nonce, BigInteger gasPrice, BigInteger gasLimit, byte[] data, int chainId);
+	Single<String> resendTransaction(Wallet from, String to, BigInteger subunitAmount, BigInteger nonce, BigInteger gasPrice, BigInteger gasLimit, byte[] data, long chainId);
 
 	void removeOverridenTransaction(Wallet wallet, String oldTxHash);
 
-    Single<ActivityMeta[]> fetchCachedTransactionMetas(Wallet wallet, List<Integer> networkFilters, long fetchTime, int fetchLimit);
-	Single<ActivityMeta[]> fetchCachedTransactionMetas(Wallet wallet, int chainId, String tokenAddress, int historyCount);
-	Single<ActivityMeta[]> fetchEventMetas(Wallet wallet, List<Integer> networkFilters);
+    Single<ActivityMeta[]> fetchCachedTransactionMetas(Wallet wallet, List<Long> networkFilters, long fetchTime, int fetchLimit);
+	Single<ActivityMeta[]> fetchCachedTransactionMetas(Wallet wallet, long chainId, String tokenAddress, int historyCount);
+	Single<ActivityMeta[]> fetchEventMetas(Wallet wallet, List<Long> networkFilters);
 
 	Realm getRealmInstance(Wallet wallet);
